LayoutTests/fast/regions/webkit-named-flow-event-auto-height-expected.txt

Tests the regionOversetChange event for regions with height:auto

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Flowing content into regions...
PASS event.target.name is "article"
PASS eventFiredCount is 1
PASS successfullyParsed is true

TEST COMPLETE
 
